dapp/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
chardet==3.0.4
cython==3.0.0
cytoolz==0.12.2
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.23.0
pycryptodome==3.17
pyrsistent==0.19.3
regex==2023.5.5
rlp==3.0.0
six==1.16.0
toolz==0.12.0
web3==6.3.0
websockets==11.0.3
yarl==1.9.2

#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --find-links=.
#
--find-links https://felipefg.github.io/pip-wheels-riscv/wheels/

certifi==2023.7.22

charset-normalizer==3.2.0

idna==3.4

numpy==1.24.3

tflite-runtime==2.12.0

pillow==9.5.0

requests==2.31.0

urllib3==2.0.4

scipy==1.10.1

scikit-learn==1.2.2